fix(NanoSpec): Add static_cast for signed/unsigned comparison (C4018)
- Fixed 1 C4018 warning in BCMenu.cpp line 1951
- Added static_cast<UINT>() for GetMenuItemCount() return value
- Ensures type safety in loop condition

+++ SRC/NanoSpec/BCMenu.cpp
@@ -1948,7 +1948,7 @@ void BCMenu::SynchronizeMenu(void)
 	UINT submenu,nID=0,state,j;
 
 	InitializeMenuList(0);
-	for(j=0;j<GetMenuItemCount();++j){
+	for(j=0;j<static_cast<UINT>(GetMenuItemCount());++j){
 		mdata=NULL;
 		state=GetMenuState(j,MF_BYPOSITION);
 		if(state&MF_POPUP){
